Swap argc and argv in the invoke interface. Also add a 1st argument representing the object, with the current intention to enable closures which need to access the captured closure environment in the call thunk.

+++ tinyclj/Runtime.h
@@ -11,7 +11,7 @@
 
 class Runtime {
     std::unique_ptr<llvm::orc::LLJIT> m_JIT;
-    std::atomic<int64_t> m_IdCounter = 0;
+    std::atomic<size_t> m_IdCounter = 0;
     std::unordered_map<std::string, TCVar *> m_GlobalVarStorage;
 
     static std::unique_ptr<llvm::orc::LLJIT> createJIT();
